requirements.txt
Flask
Flask-RESTful
Flask-CORS
Flask-SQLAlchemy
Flask-JWT-Extended
mysqlclient  
python-dotenv 
requests 